Update the-3d-window.md
There is no "Alt" key on a Mac. Pressing the control key allows to change from rotating the scene to translating the camera. This is not really necessary, as translation can also be achieved by pressing the pad with two fingers.

I'm able to zoom by pinching, but I'm not able to zoom out. None of the other Mac keys (option, command) help with this..