Update session_server.py
Just tried to create a class for global variables and so as to eliminate the use of globals in the code, although left some of them like(LOAD_THRESHOLD, LOAD_REFRESH), which are mainly constants and therefore won't introduce bugs. This is not the full implementation of the code, there are somethings that I don't know like tornado, thus don't know whether adding arguments to the SessionHandler, MonitorHandler, ClientWebSocketHandler 's __init__ will affect the working of the code, any help would be appreciated. 
I Will take feedbacks positively, even if the changes are not necessary then I can focus on other bugs else if I am going in the right direction then maybe will try to make it more robust and try to implement the same for other files as well where global is used.