Updating Darwin-opHinge2.proto with virtual RoboCup procedure
The main parameters of the joints and motors are now accessible using local
variables (as it was done for backlash).

This change has the following impacts for all joints:

- maxTorque is slightly decrease (2.5 -> 2.3)
- maxVelocity is strongly decreased (12.26 -> 5.24)
- dampingConstant is strongly increased (0.002 -> 0.27)
- staticFriction is strongly increased (0.025 -> 0.80)
- resolution of the sensor is now limited

Impact of the behavior has been tested in the following way: on a RoboCup field,
the robot can still walk and stand-up from both side.