Automated testing: support for state based time event + split of 1st scenario
This commit mainly introduces the notion of state based time event. This allows
to specify the time since we entered in a specific game state, thus making it
easier to provide explicit timing.

To make testing faster, the test of the 2 minutes elapsing before the start of
the game has been extracted to another scenario `test_initial_time`, this
reduces time to run `test_initial_positions` from around 5 minutes to 20
seconds.

The `test_initial_positions` still has many tests failing (6 or 7 depending on
runs). This is mainly due to two points:

1. The communication between the AutoReferee and the GameController seems to be
   pretty slow. A consequence is that often player are not penalized 2 seconds
   after the moment they committed an infraction. Sometimes, state is not
   changed quickly enough.
2. Blue player 1 is penalized and receives a yellow card for illegally entering
   the field. This is incorrect since it is not penalized and state is READY
   when it enters the field.