Automated testing: dropped ball scenario + features for rc_testing
Implementing an automated test scenario for the dropped ball scenario. Since the
dropped ball does not seems to be enabled actually, testing the whole scenario
requires to perform currently two actions that the AutoRef should do
automatically:
1. Applying a Dropped Ball on the GameController whend 7:48 remains for first
   half.
2. Switching state SET to PLAY after 5 simulated seconds of the SET from the
   dropped ball phase

Note that a bug appears: ball goes through BLUE 1 at the beginning of the
scenario.

In order to obtain this automatic scenario, a few features were added to
rc_testing:

- Testing which team has KickOff
- Adding support for specifying TimePoint and TimeInterval for the number of
  times we entered in a given state.
  - Some actions are triggered once READY has been reached for the second time
- Adding the notion of critical test (execution of the scenario is stopped if
  a critical test fails)
  - When Dropped Ball is not called, some events are unreacheable.