Refactor of the RoboCup 'player' + making it non-blocking
In previous version, reading on the client was blocking, which was significantly
slowing down the simulation and was not what was initially planned.

Refactoring the code into more separated function to provide more elaborate
reading.

With this major refactor, there is also the following aspects to be taken into
account:

- Adding a benchmarkLevel to the Player to have a better control on performance
  issues.
- The 'encode' part for the camera was also removed to ensure no time is wasted

The performance gain measured on a computer with GTX650 and i5-10600K, on a
simple scenario with 1 client requesting images at every timestep is the
following:

- Old player: ~0.35 speed factor
- New player: ~0.70 speed factor

NOTE: when a client disconnect, the camera stays activated, that might be an
unwanted behavior.