Allowing goals to be scored directly from FREEKICK and PENALTYKICK
The previous implementation did not properly set the `game.can_score` variable.

Adding a test scenario to check that the fix works properly.

Initially, the scenario was passing in both cases, which was surprising. After a
more thorough investigation, the source was that `ball_touched_by_opponent` was
always considered as True when `game.ball_last_touch_team` was `None`.

With the current version, test is failing if we don't set the can_score
properly.

All tests are passing.