Fix incorrect documentation of RangeFinder maxRange field
Describe the Bug
RangeFinder sets range above maxRange to maxRange instead of infinity, as mentioned in the documentation. Please update the documentation.

The maxRange defines the distance between the range-finder and the far clipping plane of the OpenGL view frustum. This field defines the maximum range that a range-finder can achieve and so the maximum possible value of the range image (in meter). If the depth is bigger than the maxRange value then infinity is returned.

Steps to Reproduce
Check rangefinder image when distances above maxRange occur in scene

Expected behavior
Documentation should be updated to the following:

The maxRange defines the distance between the range-finder and the far clipping plane of the OpenGL view frustum. This field defines the maximum range that a range-finder can achieve and so the maximum possible value of the range image (in meter). If the depth is bigger than the maxRange value then maxRange is returned.