Improvements on CMake template
Fixed a linking issue that may appear in some examples (I found it in the the crazyflie) in the CMake template.
The issue was a missing link to libm (set (LIBRARIES m...). It will show the following linking error: undefined reference to symbol 'fmax@@glibc_2.2.5'
Also added the WEBOTS_HOME variable in the CMake, just to avoid the hassle of having to export it in the console every time.