Stop pending animation on close
If an animation recording is started and Webots is closed before it is stopped, the json animation file is not created. This adds a small check to stop the animation correctly when Webots is closed by the user in the GUI, with ctrl-C in the console or by a supervisor.