Log test suite performance.
Log performance to performance_log.txt.

Change Command python class to take the command as an array instead of a
string and update code which uses it. This was needed to ensure a path
to performance_log.txt that contained spaces would work properly.

Fix syntax errors in child.py that is created by command_test.py.