Fix unitialized sliding friction when using asymmetric rolling friction.
Also update expected values in test to reflect the result of the change.

Previously mu2 was initialized when using asymmetric rolling friction, but ODE will use asymmetric sliding friction if we set `dContactAxisDep` (because `dContactAxisDep  == dContactMu2`), so we need to make sure `mu2` is set. The unitialized mu2 was resulting in the rolling_friction test failing under some circumstances.